==> REST[ES]_donors/Donor1ED801Donor11ED8/1_Load_Review_for_D1D11CDD.Rmd <==
---
title: "Load & Review Data v2 (get all Donor sorted)"
author: "Wesley"
date: "2022/03/28"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load all of the Donor Data

To start, we need to load all of the Excel files. To do this, we will use the
readxl package. In addition, we're going to need to manipulate strings in 
the following, so let's pre-emptively load that package.

```{r libraries, error = FALSE, message = FALSE, warning = FALSE}
library("readxl")
library("stringr")
```


```{r load_data}
sheets <- excel_sheets("/home/wburr/Rushali_REST/REST/Donor1Donor11_DogTrial.xlsx")  # note: spaces in names
D1D11 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_REST/REST/Donor1Donor11_DogTrial.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(D1D11) <- sheets
```


```{r merge_sheets}
D1D11m <- do.call("rbind", D1D11)
names(D1D11m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs", "1stRT","2ndRT","Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```


```{r split_RTs}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(D1D11m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(D1D11m$RTs, ","), FUN = "[[", 2)
         )
       )
D1D11m <- data.frame(D1D11m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples}
unique(D1D11m$Sample)
save(file = "/home/wburr/Rushali_REST/REST/Donor1Donor11_DogTrial.rda", D1D11m)
```
